create components navbar, main section and footer
